phenoRecFromPop function

function to make a tibble to be added to records

phenoRecFromPop(pop, bsp, stage, checks = FALSE)

Arguments

pop
The population from which to extract phenotypic records. Has to have been phenotyped

bsp
The breeding scheme parameter list

stage
At what stage of the breeding scheme this population was phenotyped. Necessary to determine the error variance and degree of replication

checks
Whether this was a population of experimentals or checks. Necessary to determine the degree of replication

Value

A tibble of phenotypic records with six columns: 1. The id, 2. The id of the mother, 3. The id of the father, 4. The name of the stage, 5. The actual phenotype, 6. The error variance of the phenotype

Details

The tibbles coming from this function will be incorporated into records useful for maintaining the phenotypic observations across years and stages. For analysis, you need these phenotypes coupled to their ids and error variances

Examples

phenoDF <- phenoRecFromPop(pop, bsp, stage)

#> Error in if_else(checks, bsp$chkReps[stage], as.numeric(bsp$nReps[stage])): object 'bsp' not found