specifyBSP function

Function to manually specify a breeding scheme parameters (bsp) object in R, rather than using a control file. Currently does not handle costs. For costs (for now), run specifyCosts() on this. Ideally this is useful for programmatically varying breeding schemes.

specifyBSP(
  schemeDF,
  nParents,
  nCrosses,
  nProgeny,
  useOptContrib = FALSE,
  nCandOptCont = NULL,
  targetEffPopSize = NULL,
  useCurrentPhenoTrain = TRUE,
  nCyclesToKeepRecords,
  selCritPipeAdv,
  selCritPopImprov,
  nChr,
  effPopSize,
  segSites,
  nQTL,
  nSNP,
  genVar,
  gxeVar,
  meanDD,
  varDD
)

Arguments

schemeDF
data.frame columns: stageNames, nReps, nLocs, nChks, nEntries, entryToChkRatio, errVars

nParents
integer number of parents to cross

nCrosses
integer how many crosses to make

nProgeny
integer how many progeny per cross

useOptContrib
logical whether to use optimal contributions

nCandOptCont
integer how many candidates to consider for opt contrib

targetEffPopSize
numeric target effective population size for OC

useCurrentPhenoTrain
logical whether to use phenotypes for parent sel

nCyclesToKeepRecords
integer eliminate data on cycles above this num

selCritPipeAdv
function used to determine selection criterion for pipe

selCritPopImprov
function used to determine sel crit for pop improv

nChr
integer number of chromosomes for the species

effPopSize
numeric historic effective population size for the species

segSites
integer number of sites segregating per chromosome

nQTL
integer number of loci affecting the trait per chromosome

nSNP
integer number of observed SNPs per chromosome

genVar
numeric genetic variance of the founders

gxeVar
numeric genotype by environment variance of the founders

meanDD
numeric mean dominance deviation. Set to zero for additive

varDD
numeric variance across loci of their dominance deviation

Value

a named list of of the parameters to specify a breeding scheme simulation

Details

All arguments are exactly as specified in the control files. Main exception is schemeDF, which is just a tibble() or data.frame version of the set of bsp arguments which are vectors (giving values for each breeding stage). Columns must have names exactly as in the corresponding arguments in control file: stageNames, nReps, nLocs, nChks, nEntries, entryToChkRatio, errVars

Examples

schemeDF <- tibble(stageNames=c("SDN", "CET", "PYT"),
                 nReps=c(1, 1, 2),
                 nLocs=c(1, 1, 2),
                 nChks=c(1, 1, 2),
                 nEntries=c(100, 50, 20),
                 entryToChkRatio=c(50, 20, 10),
                 errVars=c(150,75,40))
bsp <- specifyBSP(schemeDF,nParents = 10, nCrosses = 10, nProgeny = 10,
                useOptContrib = FALSE,
                useCurrentPhenoTrain = TRUE,
                nCyclesToKeepRecords = 1,
                selCritPipeAdv = selCritGRM,
                selCritPopImprov = selCritGRM,
                nChr = 2,effPopSize = 50,
                segSites = 100, nQTL = 5, nSNP = 10,
                genVar = 50, gxeVar = 0, meanDD = 0.05, varDD = 0.25)

#> Error in `names <- `(., bspNew$stageNames): could not find function "names <- "

test <- runBreedingScheme(replication = 1,nCycles = 1,
                        initializeFunc = initFuncADChk,
                        productPipeline = prodPipeFncChk,
                        populationImprovement = popImprov1Cyc,
                        bsp = bsp)

#> ****** 1 
#> [1] "initFuncADChk deprecated. Please use initializeScheme"

#> Error in initializeScheme(bsp): object 'bsp' not found

#> No traceback available 
#> NULL

#> Warning: restarting interrupted promise evaluation

#> Error in mget(ls()): object 'bsp' not found